🚚 created separate file for the AI
