+++ resources/js/Pages/Students/Level.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Head } from '@inertiajs/inertia-react';
 import { usePage } from "@inertiajs/inertia-react";
-import AICanvas from "@/Components/Student/AICanvas";
+import AICanvas from "@/Components/Student/AICanvas/AICanvas";
 
 import "../../../css/pages/Students/Level.css";
